tests: fix resource types in XglShaderObj
XGL_OBJECT was used as void *.  That was confusing.  Make the resource types
match what are in XglDescriptorSetObj.

Change XGL_SAMPLER* to XglSamplerObj*.  We could as well change XGL_SAMPLER*
to XGL_SAMPLER, which is itself a pointer.  But XglSamplerObj* is easier to
work with when we switch to the new C++ binding.