layers: Use objects dispatch table for key to layer maps
Since the object may vary but the dispatch table may be same, use
dispatch table as the key to various table maps rather than emplace
method to add new map key. This eliminates the need for layers that support
multiple devices  or instances to add entrypoints on the object creation calls.