tests: Changes to use new WSI swapchain extensions.
There are still some FIXME's that must be addressed, as well as fully calling
the new entrypoints.  This patch serves as a basis for review and further work
with other Vulkan engineers.

Make space in std::vector using resize.  Need vector to actually allocate space
for the number of expected data elements. Reserve doesn't do it, but resize
does.

Update testframework for new WSI extensions