Swapchain: Use an INFO instead of an ERROR for when pNext is non-NULL.
This is a compromise.  Currently, this is an error situation.  However, another
extension could be added that uses pNext, and then if the Swapchain layer isn't
updated, a false-error situation could occur.  By using an INFO message, the
user is still alerted of the possibility of a problem, but won't be annoyed in
the case of a new extension.