Remove faulty LineLength .clang-format directive.
The LineLength: 132 directive was making some clang-format
implementations stop working.  The correct directive is
ColumnLimit.

We also had agreed that the ColumnLimit would be 80 for
all code in this repo except for layers. So remove the
directive here to keep 80 as the default and to reflect
current formatting.

The layers folder has a .clang-format file that correctly
specifies ColumnLimit: 132.

Change-Id: Ic7c2a575e7d52141e3ecbc2e3644210cd85e5ac9