MR 289: layers: Change the test about "too many images".
Given new spec language coming out in early April (clarifying various aspects
of acquiring presentable images), a change was needed to the WSI-swapchain
validation layer.  The new validate-layer test will work for "BlitModel"
drivers that advertise a minImageCount of 1.  Failing the test now results in
an error instead of a performance warning.