python: Update script to include all platforms when generating ifdef code
This commit cleans up the platform detection code that was emitting some
(but not all) ifdef wrappers around platform specific code.  It changes
the logic to wrap all the functions that are platform specific.  This
allows us to use these generated files on multiple platforms, not just
the platform is was generated on.