tests: ensure WSI extensions enabled before functions called.
This is the initial version of the EnableWsiBeforeUse test case.  It verifies
that the swapchain validation layer issues an error if any of the WSI functions
are called when the corresponding extension isn't enabled.  This Initial
version tests the functions for the following extensions:

- VK_KHR_xcb_surface
- VK_KHR_surface
- VK_KHR_swapchain

There is curently one exception: vkQueuePresentKHR().  This function dispatches
from a VkQueue.  The layer uses the swapchain parameter in order to look up the
VkDevice that the extension should have been enabled for.  Since we didn't
create a real swapchain, the lookup doesn't occur (the layer doesn't chase NULL
pointers), and so the error doesn't occur.

This test case still needs to test the following extensions:

- VK_KHR_android_surface
- VK_KHR_mir_surface
- VK_KHR_wayland_surface
- VK_KHR_win32_surface
- VK_KHR_xlib_surface

When the swapchain layer supports the following extensions, this test case
should test them:

- VK_KHR_display
- VK_KHR_display_swapchain