tests: Adjust VkLayerTest.RequiredParameter
Replace vkEnumeratePhysicalDevices call with a call to
vkGetPhysicalDeviceQueueFamilyProperties in the RequiredParameter
layer test.  This test specifes NULL for a required parameter,
expecting the parameter_validation layer to report an error.  A
recent driver update on some platforms has installed an implicit
layer in front of parameter_validation, which crashes in
vkEnumeratePhysicalDevices due to the NULL parameter.

Change-Id: I46ebdec4d87b5b0039a9e39adf0e646d591f48cf