tests: Disable BindImageInvalidMemoryType test
This test causes a crash on at least one android gfx driver and it's not clear
that the way it allocates memory is within spec. Commented the test to note an
allternative method that can be used to get the desired result but also make
sure that memory is allocated in a device-supported manner.