layers: VT59, Change query count errors to warnings
In patterns where an API call is normally made twice, the first with
a NULL pointer returning the count of structures to receive followed
by second call with a pointer to the appropriately-sized buffer, the
swapchain layer was returning validation errors if the size call was
made. However, this is legal behavior so these now return validation
warnings. Also updated the relevant validation tests.

Change-Id: I9f493b27bf4fa0bec6385499c76c5cece1ad86b9