loader: Rename icd and physical device vars
Clean up the names of the icd and physical device structs, lists, and
variables used throughout the loader.  Before, it was hard to tell where
the item was to be used appropriately.  Now, the names include a hint as
to where they are expected to be allocated, freed, and used.

Also, fixed a bug where we were using a trampoline item in a terminator
function, which caused issues.  Thanks to Piers @ Nvidia for discovering
this.

Change-Id: If98628a5496e9f645eff94b73e80ae7f8408f7b4