tests: Update VU database for obj tracker tests
Change the testname column from Unknown to None for all
error enums referenced in object tracker.  This is a good
starting point because there are vitually no tests explicitly
written to test object tracker checks.
A later commit will update the few rows in the database
with the names of the tests that currently exist.

Change-Id: Ifffb5ee09a74e8e2b7b8b2e70e7e9f9b4d12afbc