layers:Mark latest buffers and images as written
Fixes #1311

In the Draw-time refactor the marking of store images and buffers as
written was incorrectly moved before the list of images and buffers was
updated for the current active bindings.

This fix restores correct order of operations so that the buffers and
images of interest are first added to cmd buffer active list, then
they're appropriately updated as written.