tests: Add loader test to catch recent bug
Added a loader test to catch the case where the user just calls
vkEnumeratePhysicalDevices with the full count and array passed
in without previously querying the count.  Also, verified that
the count returned with that call matches the count returned by
the standard call.  Finally, added a check for the VK_INCOMPLETE
return value when not enough physical devices are indicated on
a second call.

Change-Id: Ic14a109ca59dcba3ea298ec3c0092403026e9b0c