loader: Update meta-layers
Use the meta-layer Manifest API version to check against each
of its component layer API versions.  This will force us to
make sure each JSON file is properly updated in sync.

Add a section to the LoaderAndLayerInterface markdown doc about
the usage of meta-layers, and an example of how they look in a
Manifest JSON file.

Change-Id: I667cfd3013cbcf201c83935dc6f1c2bb937c7742