build: Change AppVeyor config to use multiple jobs.
This is a complete rewrite of the AppVeyor config file.

A separate job is used for each config in the Platform-Configuration
matrix.  This helps get around the problem of the 60-minute limit per job.

The invocation of the update_external_sources script has been
replaced with individual commands to build glslang and spirv-tools
only for the platform configuration selected for each job.
This avoids having to build glslang and spirv-tools 4 times for each job.