tests:Remove functional check from ExternalMemory
The ExternalMemory test is a positive test that would only pass if
underlying device functionality was verified.
Remove the functional check as that is not appropriate for LVL tests.

The purpose of positive tests in LVL is to stress the validation layers
in specific cases that are known to have had bugs or have potential for
bugs. These tests shouldn't be validating underlying device
capabilities. CTS is the appropriate place for device functional tests.