tests: Fix InvalidBarriers test
Fixed a few problems with this test to get it running on our device
profiles without errors. Firstly, fixed the logic for searching for a
queue that supports graphics but not compute. Secondly, fixed the call
to vkCmdWaitEvents so that it produces the error expected in the test.
Thirdly, added a destroy event call to the end of the test to prevent
and unexpected error. Finally, added a break statement to the for loop
that sets the test framework's device's default graphics queue family.
This is necessary as the rest of the test framework assumes that the
default graphics queue is the one found at the lowest index.

Also, added a queue to our device profiles that supports graphics and
not compute as is required by this test. In doing so, I renamed the
device profile to a generic name, and updated the info in the profile
now that the file no longer reflects the capabilities of the given
chip.