tests: Extend the device profile layer
Extended the device profile layer so that every validation layer test
can be run without skipping on just the device profile layer and the
mock ICD. This provides a quicker and more reliable way to run all the
tests than using devsim with the 9 current device profiles. This form of
running tests however is possibly less thorough as each test will only
encounter one config instead of the 9.