Add validation for VK_KHR_draw_indirect_count
This MR adds basic validation for the VK_KHR_draw_indirect_count
extension.

* Of the 35 explicit VUs added to each function, there is only coverage
for about 8 VUs. This is not different to the existing coverage for the
Vulkan 1.0 core vkCmdDrawIndirect and vkCmdDrawIndexedIndirect functions
(see
https://github.com/KhronosGroup/Vulkan-LoaderAndValidationLayers/issues/2448).
I don't know why so many of the VUs have no validation coverage. I'm
sure that part of the reason is that VUs that deal with parameters
inside the VkDrawIndirectCommand and VkDrawIndexedIndirectCommand data
in the indirect draw buffers are quite hard to implement validation for.
I have not attempted to fill in all the missing validation. It's already
marked in the validation database (vk_validation_error_database.txt) as
not implemented.

* Of those 8 VUs that have coverage, the existing validation layer test
code doesn't test all of them. Only about 5 are tested. There is no easy
test framework that can be reused to implement these tests. In fact,
looking at the existing test code there is very little test coverage for
vkCmdDraw* commands. I didn't attempt to plug those holes.

Change-Id: I225dfd08e28eaf778d9fd89b7807e9e1c8ad074a