build: Copy loader DLL only when ldr repo supplied
A previous commit added code to copy the loader DLL to the
test application binary directory so that the application can find
a loader.  This works fine when the VULKAN_LOADER_INSTALL_DIR CMake
option is used, as the application is linked with the loader export
lib from this INSTALL_DIR and the DLL is copied from there as well.

But if VULKAN_LOADER_INSTALL_DIR is not specified, the application
is linked with an export lib from an SDK (if present) and the DLL is
copied not from the SDK but from the system install dir.  This is
silly since the DLL would be loaded from the system install dir
anyway.

Change the code to copy the DLL only when VULKAN_LOADER_INSTALL_DIR
is supplied.