ci: Use update_deps script in Travis-CI
Using the script causes CI to build the "known-good" versions of the
dependent repositories. This should improve the stability of the
CI testing of Vulkan-ValidationLayers.

If Travis CI is triggered by a cron job, it builds the master branch
of the dependent repositories.

Add Vulkan-Tools to "ci_only" repo list in known-good file since
the CI tests need the mock ICD.