ci: Turn on ccache for Travis-CI
- Although ccache is turned on in the Travis config file, additional
steps are needed to turn it on for CMake-based builds.  Set USE_CCACHE=ON
in the Travis config for this repo to cause CMake to "wrap" (prepend)
compiler invocations with "ccache".
- Turn off ccache linker wrapping in CMake code to enable ccache.
It doesn't work.
- Modify the helper generator script to sort the enums when generating
the enum-to-string functions.  This causes the header file to be generated
with the same content each time the script is run, which greatly improves
ccache effectiveness.  For local ccache-assisted builds of this repo,
the build time can be cut in half.