layers: Use pointer-to-member in shader caps table
The shader capabilities table was storing pointers to specific enable
booleans, but the table has a static lifetime so only capabilities from
the first VkDevice were being captured and used for comparison.  This
change goes back to storing pointers-to-member so that the capabilities
used for comparison reflect the current device and the lookup table can
remain static.

A wrapper struct is used for feature pointers because there are multiple
feature structs. The wrapper applies each pointer-to-member to the
correct feature sub-struct based on its type.