layers: Validate shader stage input/output limits
This addition adds validation for the following shader stage limits:
  - maxVertexOutputComponents
  - maxGeometryInputComponents
  - maxGeometryOutputComponents
  - maxTessellationControlPerVertexInputComponents
  - maxTessellationControlPerVertexOutputComponents
  - maxTessellationEvaluationInputComponents
  - maxTessellationEvaluationInputComponents
  - maxFragmentInputComponents
These are all specified in the VkPhysicalDeviceLimits structure.

These checks do only adhere to the Component-decoration
side of these limits. I.e. if maxVertexOutputComponents=128, it's not
possible to pass 128 individual floats (each with its own location), as
there is also a Location-decoration limit:
maxVertexOutputComponents/4=32. Yet, these 128 floats can be specified
to occupy different components of these 32 locations using the
Component-decoration. Thus, the validation performed by this addition
is still valid, even though it does not consider the number of
locations used. (See section 14.1.14/15 of the spec)

A float that is specified to be 64-bit, counts as 2 components,
while all 32-bit (or less) count as 1.

Precision qualifiers are disregarded as SPIR-V spec 2.14 states
"All loads and stores involving relaxed precision still
read and write 32 bits of data, respectively.".

Calculate numComp using built-in decorations

- All checks are performed by checking whether or not a variable is
  built-in. If it is, then the components for said variable
  (the variable as a whole) are disregarded.
- Added back tests for tessellation stages.
- All tests passed locally, and clang-format has been applied,
  so Travis should pass.

Add tests for all 4 shader stages

Tests exceeding the shader stage's input/output limits have been added.
As the limits can vary between devices, the shaders are dynamic in
the sense that they can vary from device to device.
There are 5 new tests:
  - Vertex output
  - Tessellation Control input and output
  - Tessellation Evaluation input and output
  - Geometry input and output
  - Fragment output

The number of components is calculated using built-in decorations

- All checks are performed by checking whether or not a variable is
  built-in. If it is, then the components for said variable
  (the variable as a whole) are disregarded.
- Added back tests for tessellation stages.
- All tests passed locally, and clang-format has been applied,
  so Travis should pass.