layers: Add check and test for relax block layout
Tell the SPIR-V validator to skip block layout checking if the
VK_KHR_relaxed_block_layout extension is enabled.
This prevents validation error messages for shaders with
non-compliant layouts when the extension is enabled.

The VK_KHR_relaxed_block_layout extension was promoted to core
in 1.1, so it is "enabled" for 1.1 and later apps all the time.

Since block layout checking is always relaxed in 1.1 or later,
there is really no way to write a negative test for 1.1 or later.
Therefore there is just a positive test that checks that no error
is generated for a shader with a non-compiliant layout.

Fixes #275