layers: Fix object tracking of VkDisplayKHR handles
Applications using Direct-to-Display are asserting when unique
object tracking validation is enabled due to bugs handling the
VkDisplayKHR handles.

There is no single point of creation for the display handle
when running direct-to-display; a number of APIs return the
same display handles. The existing code ends up wrapping the
same VkDisplayKHR in multiple new unique_ids. The assert occurs
because GetDisplayPlaneSupportedDisplaysKHR() tries to lookup a
handle in the map that doesn't exist. There is a TODO in that
function that says "What's going on here?"

This fix first checks the known display handles and returns a
match if there is one before resorting to creating new mappings.