build: CMake enhancements for unified Vulkan build
Uses imported targets for external dependencies whenever possible.

Note that Gtest is built statically under the current unified Vulkan
build, so we add checks to adjust accordingly based on a global variable
set by the parent build in the unified Vulkan. This should be revisited
later on.

Namespaced the internal build targets to avoid naming conflicts.

Change-Id: I7c855cf44c01b5f0620a86975c6f91e1fc5b9cb8