layers: replace VUID error map with struct
Replace the std::unordered_map container holding the VUID -> spec text
mapping with a vanilla struct of char * literal strings. The map is
now built on first use from the string table.

This change is a workaround for slow build times observed with use of
the directly initialized unordered_map, in particular on 32-bit Android
builds.

Change-Id: Ic0ab0bf443a57ee616b3cac17fc8ea0442e6d0c2