build: Update known-good for 1.1.91 header
This commit includes all necessary changes to the ValidationLayers
repository for updating the known-good to the 1.1.91 header. Included in
this update:
- Rename promoted NVX extensions to NV - Names of functions, enum
values, etc. related to NVX raytracing were changed to NV ray tracing
    Updated:
    - `layers/core_validation.cpp`
    - `layers/core_validation_types.h`
    - `layers/descriptor_sets.cpp`
    - `layers/shader_validation.cpp`
    - `layers/shader_validation.h`
    - `layers/threading.h`
- Fix failing test for 1.1.91 header update - VkLayerTest
`UnrecognizedValueBadFlag` was changed to meet the requirement that if
`waitSemaphoreCount` is greater than 0, `pWaitSemaphores` is a valid
pointer to an array of valid `VkSemaphore` handles and redfines a "bad
flag" due to the addition of pipeline stage flag bit
`VK_PIPELINE_STAGE_ACCELERATION_STRUCTURE_BUILD_BIT_NV`
    Updated:
    - `tests/layer_validation_tests.cpp`
- Propagate `genvk.py` changes from 1.1.91 - Changes to Vulkan-Headers's
`genvk.py` were added to Vulkan-ValidationLayers's `lvl_genvk.py`
    Updated:
    - `scripts/lvl_genvk.py`
- Generate a new `layers/vk_validation_error_messages.h`
    Updated:
    - `layers/vk_validation_error_messages.h`
- Update known-good for 1.1.91 header
    Updated:
    - `build-android/known_good.json`
    - `scripts/known_good.json`