layers: Added VK_NV_framebuffer_mixed_samples
Added support for VK_NV_framebuffer_mixed_samples to the validation
layers.

Test the following VUIDs in case that the extension is present:
VUID-VkGraphicsPipelineCreateInfo-subpass-01411
VUID-VkGraphicsPipelineCreateInfo-subpass-01412
VUID-VkPipelineMultisampleStateCreateInfo-rasterizationSamples-01415
VUID-VkPipelineCoverageModulationStateCreateInfoNV-
coverageModulationTableEnable-01405

Test the following VUIDs in case that the extension is not present:
VUID-VkGraphicsPipelineCreateInfo-subpass-00757

Added tests FramebufferMixedSamplesNV and FramebufferMixedSamples
to verify that the relevant VUIDs are touched.