gpu: Add missing stride decoration
This is a temporary fix to address a missing array stride
decoration in the SPIR-V generated by the instrumentation pass.

This fix should be removed once a spirv-tools version with the fix
is available via known-good.  This fix is designed to coexist with
a fixed spirv-tools, so it does not have to be removed immediately
when using the fixed spirv-tools.