vk_validation_stats: Apply a bit of pretty-printing to the header
One of the challenges of hooking up new validation layer tests is the
difficulty in finding the VUs that are wanted, and then copy/pasting the
correct VUIDs into the source tree. To aid in this process, sort VUIDs as best we can by
extension, based on a few hand-tuned heuristics.

This should not change any output, it just makes this file and the
validation layer codebases a bit easier to read and maintain.