tests: Fix test failures on maxPushDescriptors 0
Fixed test failures and assertions when the maxPushDescriptors property
was zero on platforms that support the push descriptor extension.

Additionally skipped test cases with "unexpected" errors with the same
issue.

Note: it *is* invalid for maxPushDescriptors to be zero when the
extension is supported, but as the platform on which this occurs is
common, the CI team preferred the work-around to the false positives.

Change-Id: Ibf366645562d95442526564587786331d3335586