Revert "build: Fix test build to use loader as imported target"
This reverts commit c0269834f658b68b6f9f878f2d9bba6d71b4301d.

The newly added line 108:
    target_include_directories(vulkan INTERFACE "${Vulkan_INCLUDE_DIR}")
is breaking CMake 3.5.1 for reason yet to be determined.

Temporarily reverting this commit until the root cause is identified and
fixed.

The breakage seems related to the version of CMake.

CMake 3.5.1 is the version provided by Ubuntu 16.04 LTS "Xenial",
which is an officially supported Linux version.

The breakage was not discovered by GitHub's Travis CI because the Travis
definition of "Xenial" unexpectedly provides CMake 3.12.4