layers: Add plane aspect bits as valid flags
According to the description of vkCmdCopyBufferToImage: "
If the format of dstImage is a multi-planar image format,
regions of each plane to be a target of a copy must be specified
separately using the pRegions member of the VkBufferImageCopy
structure. In this case, the aspectMask of imageSubresource must
be VK_IMAGE_ASPECT_PLANE_0_BIT, VK_IMAGE_ASPECT_PLANE_1_BIT, or
VK_IMAGE_ASPECT_PLANE_2_BIT.". This is relevant for YCbCr image
formats.

As of Vulkan 1.1.0, YCbCr is core and the mentioned aspect flag
bits should not yield validation layer errors any more if used
as the aspect flag for a region in vkCmdCopyBufferToImage. The
same holds true for vkCmdCopyImageToBuffer, vkCmdBlitImage and
vkCmdCopyImage.

This has now been added to parameter_validation_utils. A check is
performed to check if the physical device >= 1.1 support. If yes,
these aspect flag bits are taken into consideration. If not, the
old check to see if the YCbCr extension is supported is performed.