Fix check for compatible AHardwareBuffer usage bits.
The test claims that other usage bits besides the ones list in the Usage Equivalence table are not allowed to be present on the AHardwareBuffer. However in the spec it reads:

"The usage member of AHardwareBuffer_Desc must include bits corresponding to bits included in VkImageCreateInfo::usage and VkImageCreateInfo::flags where such a correspondence exists according to AHardwareBuffer Usage Equivalence. It may also include additional usage bits, including vendor-specific usages."

This suggests that other bits are allowed to be present.