perf: Optimize unordered_map usage in thread_safety.h
Wrap std::unordered_map in a 'small_unordered_map' class that has inline storage
for the first element of the map. This avoids memory allocation and hashing in the
common case where there is only zero or one entries in the map at a time, which
happens a lot given that entries only stay in the map for the duration of an entry
point.