Fixes when VkExternalFormatANDROID.externalFormat = 0
This change adds checks to more spots to see if VkExternalFormatAndroid.externalFormat = 0. If it does equal 0 then the spec states that it should be treated as if the VkExternalFormatANDROID struct was not even present. This was already correctly checked for in ValidateCreateImageANDROID, but this change adds the check to more spots in the layers.