layers: install headers in vulkan subdirectory
The header repository installs headers in
$CMAKE_INSTALL_INCLUDEDIR/vulkan [1].

The basically leaves no choice to the user, forcing the vulkan
subdirectory to be created. The validation layers don't have that same
behavior and it seems that a few Linux distribution choose a different
path that makes it impossible for applications or custom layers to
find the include path for the headers.

[1] : https://github.com/KhronosGroup/Vulkan-Headers/blob/8e2c4cd554b644592a6d904f2c8000ebbd4aa77f/CMakeLists.txt#L39

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>