tests: set target API version correctly for some tests
Both tests had been configured for Vulkan 1.0 version but
used some non-KHR functions that were not in core until 1.1,
so the respective pointer was NULL and used stubs instead,
which means than some objects were not properly initialized.

As these are tests for the validation layers themselves, the
patch sets the target version to Vulkan 1.1 to fix this.

Fixes segfaults on the following tests on Intel's ANV driver:

VkPositiveLayerTest.MultiplaneImageTests
VkLayerTest.MultiplaneImageSamplerConversionMismatch