build: Remove unneeded gtest include dir from VLT
The gtest project exports its include directories as
"INTERFACE SYSTEM" which means when including the gtest target
in another target, it is unnecessary to explictly list the
gtest include directories in the target's include directory
search path.  The target in this case is the validation layer tests (VLT).

Further, it is somewhat misleading to add the gtest include dir in the
search path on Linux-like systems because gtest exports the include
dir as "SYSTEM" which means it is moved from the location specified in the
CMake code to the end of the search path, along with other system includes.
So placing it in the include path in the CMake code gives the misleading
impression of its position in the final include search path.

Motivated by https://github.com/KhronosGroup/Vulkan-ValidationLayers/issues/821.