layers: Split parts of core_validation out
This is a bikeshed proposal. Currently, core_validation.cpp is a single
file that is over 12,000 lines long. It takes a long time to compile,
can be exhausting to navigate, and some tools like IntelliSense can have
issues with its large size. Splitting it into multiple files isn't a
panacea for the above issues, but a small step towards a larger goal.

Here, I do drawdispatch as a simple example to show the idea itself and
build approval. It was small and self-contained and relatively
conflict-free. If this is well-received, I will do the work myself to
help split core_validation.cpp into multiple more files.