layers: Enable VK_EXT_debug_utils in uber layer
In layer consolidation with the migration of the various debug
implementations there was some duplication. Remove the duplication.

Enable the debug utils extension in the uber layer.

Note: because of the need to track the lifetime/reset state of command
buffers, the vkCmd.*DebugUtilsLabel calls are only supported in when the
core validation sublayer is enabled.

Change-Id: I912cbeda27521806606677c8067c13e274e8184c