layers: portability validation layer (for review only)
DO NOT MERGE, FOR REVIEW ONLY!

Initial implementation of a portability_subset v0.2 validation layer
using the new Validation Layer Chassis architecture.

Includes a rough draft of the validation error messages to be reported,
as derived from the VK_EXTX_portability_subset.md specification.
Both the list of error messages and the contents of those messages need
review and refinement.

This frames out the layer for intercepting the necessary Vulkan calls
for implementing the validation logic.

Builds on Windows and Linux as a separate shared library.
When integrated into core Vulkan, this functionality can be incorporated
directly into the KHRONOS_validation layer.

The RUN_PORTABILITY_VALIDATION.sh script for Linux demonstrates the
layer using a modified portability-aware vkcube executable.

Due to the extension not existing in the Vulkan XML manifest, there
are several hacks to the Layer Chassis to temporarily bypass its
code-generation.

Included is a verbatim local copy of vk_extx_portability_subset.h from the
v0.2 release.

Change-Id: I6394adadb5330776110683404f74b76309a78592