layers: Fix use-after-free bug for push descriptor sets
The LAST_BOUND_STATE struct tracks the currently bound descriptor sets
in member 'boundDescriptorSets' which can point to a push descriptor
set in member 'push_descriptor_set'. The function
RecordCmdPushDescriptorSetState() was reseting the push descriptor
set without updating the pointer in the bound descriptors sets
causing a use-after-free in the lambda in
UpdateLastBoundDescriptorSets().